Created BaseClass for all salt-api related actions. refactored Runner action to use prepared request
